backend/requirements.txt
fastapi
uvicorn
python-docx
PyMuPDF
openai
langchain
faiss-cpu
python-multipart
python-dotenv
tiktoken
python-docx